feat(http+promises): partially integrated
superagent+bluebird promises

I still need to figure out the best place to use it in the context of
Redux which is what I'll have to figure out next. I think once that's
done then we're pretty much almost complete with this bp (at least the
core stuff anyway).

There are a few other nice to have components:

1. JS Standard (that badge looks cool and I want the JS style to be
consistent).
2. ReactJS hot-reloader
3. TravisCI but this means I need to integrate JEST
4. ESLint - I've tried doing this in other projects that are written in
Babel but I've had issues in the past.

Still in the TODO:
1. Redux
2. Bootstrap
3. Testing